[proxy] Move some tests to a dedicated module
This unclutters the pivotal `proxy.rs` module.
